+++ services/aws/grey-user/pkg/databases/database.go
@@ -1,4 +1,4 @@
-// Path: pkg/databases/database.go
+// Path: grey-user/pkg/databases/database.go
 package databases
 
 import (
@@ -11,5 +11,5 @@ type Database interface {
 	PutItem(ctx context.Context, item interface{}, conditionExpression *string, tableName string) error
 	GetItem(ctx context.Context, key map[string]interface{}, tableName string) (map[string]interface{}, error)
 	DeleteItem(ctx context.Context, key map[string]interface{}, tableName string) error
-	QueryItems(ctx context.Context, tableName string, limit int64, lastKey string) ([]map[string]interface{}, string, error)
+	QueryItems(ctx context.Context, tableName string, limit int32, lastKey string) ([]map[string]interface{}, string, error)
 }
